Feat: Validacoes parte 1 controle de fluxo

+++ app/Http/Controllers/MarcaController.php
@@ -35,7 +35,11 @@ class MarcaController extends Controller
 
     public function show($id)
     {
-        $marca =  $this->marca->find($id);
+        $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Recurso pesquisado nao existe'];
+        }
+
         return $marca;
     }
 
@@ -49,6 +53,10 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Impossivel realizar a atualizacao. O recurso solicitado nao existe'];
+        }
+
         $marca->update($request->all());
         return $marca;
     }
@@ -57,7 +65,11 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'O recurso solicitado nao existe'];
+        }
+
         $marca->delete();
-       return ['msg' => 'A marca foi removida com sucesso'];
+        return ['msg' => 'A marca foi removida com sucesso'];
     }
 }
